fix forms

+++ src/views/TableList/PlaygroundList.jsx
@@ -9,6 +9,8 @@ import 'bootstrap/dist/css/bootstrap.rtl.min.css';
 import headerLinksStyle from "assets/jss/material-dashboard-react/headerLinksStyle";
 import { CustomInput, IconButton as SearchButton } from "components";
 import { Search } from "@material-ui/icons";
+import { RHFInput } from 'react-hook-form-input';
+import { Controller, useForm } from "react-hook-form";
 
 const styles = theme => ({
   container: {
@@ -36,7 +38,7 @@ const TableList = ({ ...props }) => {
   const [formData, setFormData] = useState(null)
   const [open, setOpen] = useState(false);
   const [forUpdate, setForUpdate] = useState(false);
-
+  const controller = new window.AbortController();
   const loadData = (data) => {
     if (!data)
       data = []
@@ -55,15 +57,14 @@ const TableList = ({ ...props }) => {
     }
   }
 
+  const { control, setValue, register, handleSubmit, formState: { errors } } = useForm()
 
   useEffect(() => {
-    const controller = new window.AbortController();
     fetchData(controller)
     return () => { controller.abort() }
   }, [])
 
   useEffect(() => {
-    const controller = new window.AbortController();
     fetchData(controller)
     return () => { controller.abort() }
   }, [pageNumber])
@@ -76,6 +77,7 @@ const TableList = ({ ...props }) => {
   const handleClose = () => {
     setOpen(false)
     setForUpdate(false)
+    setFormData(null)
   }
   const nextPage = () => {
     setPageNumber(pageNumber + 1)
@@ -87,22 +89,39 @@ const TableList = ({ ...props }) => {
   }
 
   const handleEditClick = ({ prop, key }) => {
-    const data = playgroundsList[key]
+    let data = playgroundsList[key]
+    let _data = { ...data };
+    _data.id = data._id
+    _data.name = data.name
+    _data.description = data.description
+    _data.city = data.city
+    _data.price = data.price
+    _data.type = data.type
+    _data.dayStartTime = data.dayShift.start;
+    _data.dayEndTime = data.dayShift.end;
+    _data.nightStartTime = data.nightShift.start;
+    _data.nightEndTime = data.nightShift.end;
+    _data.services = data.services.map(x => ({ label: x.name, value: x._id }))
     setForUpdate(true)
     setOpen(true)
-    setFormData(data)
+    setFormData(_data)
   }
 
   const handleDeleteClick = ({ prop, key }) => {
-    deletePlayground(playgroundsList[key].id).then(
-      res => {
-        let list = tableRows.filter((_, i) => i !== key)
-        setTableRows(list)
+    const answer = window.confirm("هل أنت متأكد")
+    if (answer)
+      deletePlayground(playgroundsList[key].id).then(
+        res => {
+          let list = tableRows.filter((_, i) => i !== key)
+          setTableRows(list)
 
-      }
-    )
+        }
+      )
   }
-
+  const onSubmit = (data) => {
+    console.log(data)
+    getAllPlaygrounds(pageNumber, pageSize, controller, data).then(res => console.log(res))
+  }
   return (
     <div>
       <Grid container>
@@ -125,24 +144,32 @@ const TableList = ({ ...props }) => {
             cardSubtitle="من الأحدث إلى الأقدم"
             content={
               <div>
-                <CustomInput
-                  formControlProps={{
-                    className: classes.margin + " " + classes.search
-                  }}
-                  inputProps={{
-                    placeholder: "البحث بإسم الملعب",
-                    inputProps: {
-                      "aria-label": "البحث بإسم الملعب"
-                    }
-                  }}
-                />
-                <SearchButton
-                  color="white"
-                  aria-label="edit"
-                  customClass={classes.margin + " " + classes.searchButton}
-                >
-                  <Search className={classes.searchIcon} />
-                </SearchButton>
+                <form onSubmit={handleSubmit(onSubmit)} >
+                  <RHFInput
+                    as={
+                      <CustomInput
+                        formControlProps={{
+                          className: classes.margin + " " + classes.search
+                        }}
+                        inputProps={{
+                          placeholder: "البحث بإسم الملعب",
+                          inputProps: {
+                            "aria-label": "البحث بإسم الملعب"
+                          }
+                        }}
+                      />}
+                    rules={{ required: true }}
+                    name="filter"
+                    register={register}
+                    setValue={setValue} />
+                  <SearchButton
+                    color="white"
+                    aria-label="edit"
+                    customClass={classes.margin + " " + classes.searchButton}
+                  >
+                    <Search className={classes.searchIcon} />
+                  </SearchButton>
+                </form>
                 <Table
                   handleDeleteClick={handleDeleteClick}
                   handleEditClick={handleEditClick}
